pyproject.toml
[tool.poetry]
name = "gateway-sequences"
version = "0.1.0"
description = "Find all gateway sequences from SnapGene plasmid library"
authors = ["Manuel Lera-Ramirez <manulera14@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
biopython = "^1.84"
pydna = "^5.2.0"


[tool.poetry.group.playwright.dependencies]
playwright = "^1.48.0"
bs4 = "^0.0.2"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
